Screen: Print a warning if pageflipping was not used during sync tests.
This pretty much only ever applies to OSX with AMD gpu and PTBKerneldriver
loaded. That's also the only combo where this failure happens quite often
on Apples trainwreck with modern AMD gpus.

So far the message was only printed at high debug levels, but outputting
a failure notification already at regular verbosity level makes sense, to
short-cut the whole debugging process on the forum to "upgrade to Linux"
if this message shows.