Screen: Make blue-line stereo sync lines 3 scanlines high.
Instead of only one scanline. Too many drivers have off-by-one
bugs in vertical direction, which can lead to clipping the sync
lines off the bottom of the display. Or contrary, some stereo
hardware is sensitive to the exact position of the sync line.

Make this more robust by just extending the "line" to a "quad",
from specified vPos to vPos + 2.