Screen('ConfigureDisplay','Brightness') fail gracefully for pure queries.
If a pure brightness query fails due to lack of support by the system
then return a brightness of -1 for "not supported/failed" instead of
aborting with an error message. This way scripts can test the waters
with a pure brightness query before making the bold move of changing
brightness. Allows user-script to handle lack of brightness support
gracefully.

Trying to set a brightness will still fail with script abort and
error message, as no script which expects to change brightness
should just continue uninterrupted if that step does not succeed.

-> Successfully tested on Linux, OSX test pending.

-> Fixes #423