Screen/Linux: Make mmio mapping potentially more robust/streamlined.
Write access is only useful on AMD, so only try that on AMD,
but go for read-only immediately on NVidia and Intel. This
simplifies the code and may make it more robust in corner cases,
where the old code would have failed the write mapping with an
error other than EAGAIN. Unclear if such a scenario exists and
if this helps, but at least it should not hurt. See forum message
22544 for a potential use case.

Also reword some failure help text to clarify a bit.