Handle a very rare edge case
After  using this routine for over 20 years, I managed to call it in a way that didn’t work.  Can’t pass a single wavelength as just a number for either the input or output format, because then you don’t know the wavelength sampling, which is needed to do the appropriate power conversion.

Added error check for this case, along with a message as to how to call in a way that will work right.