Timeglue/Windows: Improve single-core locking on Windows.
Instead of assuming our threads can run on any core, query
the parent process's affinity mask to only request possible
cores. Also, in case of WinXP or earlier, or in troubleshooting
time-keeping problems on Vista and later, no longer force all
threads to core 0, which might be not allowed, but force them
onto the first allowable core according to the process mask.

Also add a bit more debug output in case of affinity set failure,
and suppress error messages if "locking" the thread to all cores
fails, which is a non-event.

This all to improves robustness and debugability in the very
unlikely scenario that anything would go wrong with core locking.

So far only one failure ever is known, which apparently resolved
itself (forum msg. 22596), and most likely has nothing to do with
the theoretical issues fixed in this commit, but perfectionism is
its own reward...