Screen: No longer 'dacbits' in 'ReadNormalizedGammaTable' - it's bs.
Report 0 for "I don't know" instead of the hard-coded 8 bpc for
Windows and Linux, and the value reported on macOS, which is almost
always not trustworthy non-sense, and therefore misleading.

I think it is not really possible to reliably report the actual
effective video output bit depth on modern OS'es + gpu's + displays.
There is too much variability in what which os'es with which drivers
on which hardware over which outputs to which displays support in
which configurations. And too many fallback and failure cases. None
of the systems has an even remotely reliable way of reporting back
what precision actually leaves the output.

So for now be just honest and don't report potential garbage.