Screen/Linux/X11: Detect and report if primary gpu provider is modesetting-ddx.
Screen('GetWindowInfo', win, 8) returns 1 if the X11 screen associated
with win is driven by the modesetting-ddx as primary gpu provider, 0
otherwise.

This is needed by XOrgConfCreator to make some decisions, based on
modesetting-ddx or gpu specific ddx in use.

Since about Ubuntu 16.04.3-LTS the modesetting-ddx is used by default
on Intel gpu's, instead of the intel-ddx, so we need to take this
into account vs. older or other distributions.

Uses RandR 1.4 provider queries.