Fixes to OtfToPsf and PsfToOtf
These fixes address numerical issues in fft and ifft.  We were previously trying to detect when the PSF was sufficiently symmetric that we should force the OTF to be real.  This seemed to be working great, but it turns out there are cases where the check failed and this led to PSFs with negative values and led to the two functions not self-inverting very well.

Removed this feature of PsfToOtf, there isn’t really any reason to avoid a complex OTF.  Added code that prevents negative/complex PSFs, but I don’t think that was actually necessary once complex OTFs were allowed.  But it seems like a good idea to keep.

Also fixed a buglet in PsychOpticsTest, where a conversion was being done in the wrong direction.  It was a coincidence that it came close to passing an equality check.  Fixing it improved the agreement from ~10-3 to ~10-16, which makes more sense.  Removed comment that wondered why agreement wasn’t better than it was.