Screen: Disable GstDeviceMonitor/DeviceProvider support on Linux.
GstDeviceMonitor, GstDeviceProvider and GstDevice require GStreamer-1.4.0 or later
to be installed at compile-time on the build system, and at runtime on the target
runtime system.

Disable this on Linux as of beginning of 2018, because we still want to keep
Ubuntu 14.04 LTS supported atm., which does not provide GStreamer 1.4+ runtimes
by default, so trying to run a Screen mex file built on a 1.4+ system (like
Ubuntu 16.04 LTS) would cause load-time linker failure on Ubuntu 14.04. I want
to switch my build system from 14.04 to 16.04 or 18.04 LTS soon, so this would
become relevant as a failure mode on 14.04. Another approach apart from compile
time disable on Linux would have been to use runtime on-demand loading/linking,
but that's too much hassle for the little to no gain in functionality on Linux
atm.

This is no functional loss, because so far we didn't use it on Linux at all with
no negative consequences, as building was always done on 14.04 LTS. The only video
source under Linux currently (beginning 2018) supported by GstDeviceProvider /
GstDeviceMonitor is the v4l2src source, but that source is perfectly handled by our
code since over a decade without the need for GstDeviceMonitor.

GstDeviceMonitor is more beneficial on MS-Windows and possibly macOS, where it
can make a positive impact on the reliability of video device enumeration. On
those systems, users have to install GStreamer runtimes of v1.4+  manually anyway,
so we know it works there as runtime requirements are fulfilled.

-> This has no immediate consequences, except i can now build on Ubuntu 16.04 LTS
   and later, so this is prep work for me switching to 16.04/18.04 LTS soon
   hopefully without backwards compatibility issues.