Screen for 64-Bit Octave on Linux rebuilt on a Ubuntu 16.04.3 build system.
Usually all mex files are built under 14.04.5 LTS for compatibility.
Building Screen under 16.04 LTS is a test if the backwards compat
measures work, iow. if code built on 16.04 against Octave-4.0 still
works in Octave 3.8 on 14.04 LTS. Prep work for a transition of the
main build system - and only remaining 14.04 test system - to 16.04
or even 18.04.