PsychPortAudio: Handle lifecyle issue (#123)
Existing code used audiodevicecount as the next handle, but on subsequent PsychPortAudio('Close', id) commands, audiodevicecount decrements, regardless of which handle is being closed, causing issues where streams would get overriden upon the next PsychPortAudio('Open') or PsychPortAudio('OpenSlave'). This was fixed by creating a small utility method nextHandle that looks for an audiodevice with a NULL stream and returning its index to serve as the next handle.