Screen: Fix Screen('Close', window) logic.
Closing the last onscreen window on a given screen would close
all onscreen windows on all screens! Why? Because the function
PsychIsLastOnscreenWindow() does not check if a window is the
last onscreen window, but if it is the last one on the screen
to which that window is associated.

Now we have 2 functions which do something that is more
intuitive from their naming:

PsychIsLastOnscreenWindow() checks for last onscreen window
in a session.

PsychIsLastOnscreenWindowOnScreen() checks for last onscreen
window on the screen associated with that window.

This should fix the Screen('Close', onscreenwindow) handling.
Also code formatting cleanup in SCREENClose.c.

-> Multi-X-Screen test pending.