Screen('OpenProxy'): Assign a non-empty rect to the proxy-window.
Just use copies of the rects of the parent window for simplicity,
although the actual size of the rect should not matter, 1x1 pixel
would have been fine as well.

This to avoid failures in the PsychSetDrawingTarget() -> PsychSetupView()
path due to empty rects, which would call gluOrtho2D(0,0,0,0,1,-1), which
leads to a GL_INVALID_VALUE error in glOrtho().

-> Tested, but Screen rebuilds pending for all platforms.
-> Also white-space/formatting cleanup.