PsychHID: Remove printf in CTRL+C keyboardhelper handling code.
It got called from the keyboard queue thread when CTRL+c was
pressed, even if console output wasn't blocked by using output
redirection, to give the user some feedback. There is some
indication that Matlab R2018a may not like that anymore and
crash.

As this is non-essential feedback, let's just get rid of it.