BitsPlusPlus: Improvements to 'OpenBits#' - handles, port specs.
Allow to select between multiple port spec strings in the
BitsSharpConfig file by passing a numerical line index as
portSpec, e.g., index 1 selects the port in line 1 of the
config file, index 2 == line2, ...

Return a device 'handle' for each device opened with OpenBits#,
additionally the underlying IOPort handle for external low-level
users, e.g., PsychRTBox. Currently we are still limited to one
open device handle at a time, this just establishes api for adding
multiple open devices in the future.

New syntax:

[handle, portHandle] = BitsPlusPlus('OpenBits#' [, portSpec]);

This works because in the past a return value of 0 vs. 1 meant
open failed vs. open success. Now a returned handle of 0 still
means open failed, whereas a returned handle of 1 refers to the
successfully openend 1st device. A value of 2, ... n would be
the handle for device 2, ... n, but this is not yet implemented.

-> Untested on real hardware so far.