ScriptingGlue/Python: Stop-Gap implementation of struct arrays.
It "works" as a PoC, but the wrong way around. Instead of having
an array with each slot being a dictionary ~ struct, we have a
dictionary ~ struct, with each element being an array.

Stores the same information, but the structure-of-arrays style
instead of the array-of-structures style is awkward to index and
awkward to visually look at as a human => A usability problem.

Maybe need a completely different implementation, potentially
roll-your-own, or NumPy based. For the moment this will do as
a stop-gap though.

-> Makes GetSecs.do('Version') "work" in an awkward way.