LoadIdentityClut: Make AMD gpu detection a bit more robust. Screen rebuilt.
Match on "AMD" substring instead of "on AMD", as the former works with
current Mesa 18, the latter doesn't.

-> Screen for 64-Bit Octave on Linux rebuilt for testing. This one
   handles AMD's amdgpu-kms DisplayCore correctly wrt. identity lut's
   and display detection, as now verified on a DCE-8 AMD Mullins CIK
   gpu in a dual-display config, with a Bits# for verification of
   identity pixel passthrough.

One problem remains: Testing the fallback path shows that the correct
LUT for fallback on AMD DC would be type 0, but we choose a different
type by default. Problem: Different ones needed for DC vs. non-DC and
currently no easy way to detect from within M-File if DC is active or
not. However, at the moment our low-level DEGAMMA lut bypass works fine
to solve all these problems, so the non-perfect fallback does not matter
in practice, and BitsPlusIdentityClutTest allows manual assignment of the
proper type 0 lut.