PythonGlue: Return a single-element structArray as a struct / dict.
Return a structArray with one element as a single dictionary, so
usercode/user can avoid the redundant [0] index.

This is probably not a good idea, but as it is contained in this
easily reverted commit, let's see. Many PTB functions which return
structArrays actually only return a single struct, which Matlab/Octave
handles transparently ie. myStructArray(1) is the same as myStructArray
in that case. This is convenient and would be convenient in Python as
well, so this commit approximates that. However, the downside is that
Python doesn't handle indexing into such a single dict as [0]['field'],
so we might get into trouble when a function that usually returns a
multi-element array occassionally only returns a single-element array,
which would turn into a dict with this patch. Maybe there is some easy
way to handle this with standard Python methods, so this patch would
make sense, maybe not, in which case -> revert.