PythonGlue: Fix PsychSetStructArrayNativeElement()
We can't just wrap PsychSetStructArrayStructElement(), as that expects
the to-be-added element to be a struct itself, not some other datatype.

We need to add things like integer lists or matrices in various modules,
so add essentially the implementation of PsychSetStructArrayStructElement(),
but without the check/error-handling for "added element is not a struct".