PythonInit: Add Python style help() support.
Add some copyright message to the Python-3 module spec, which
denotes the current code as not for public use, but all rights
reserved.

Add call to new helper PsychBuildSynopsisString() to build and assign
the normal module synopsis text as generated by the modules (upgraded)
InitializeSynopsis() to the main entry point (=dispatch) of the module.

This gives reasonably good help text when calling help(modulename) or
help(modulefunction) aka help(modulename.modulefunction).

Verified to stay compatible with Octave and work properly on both
Python 2.7 and Python 3.6.

The help system in general needs some rework in the future, we could
consolidate more code into shared helpers and make this more elegant
and less error prone. But this is something for the future. For now
this will do for both Octave/Matlab and Python.