PythonGlue: Do not accept a 1st argument as string which is not a subfunction.
So far, on a subfunctions enabled module, if the 1st argument was a string
and it didn't match a subfunction name, we just invoked the basefunction,
hoping that was the synopsis function.

Now do the same, but after calling he (presumed) synopsis function, abort
with "unknown subfunction name error state E", which makes way more sense,
so typos in subfunction names don't just trigger a printout of the synopsis
followed by continuation of the script and mysterious followup failures.

-> TODO: Should consider picking this for the mex api side as well for
         consistency and more sensible behavior.