ScriptingGlue: Rid PsychAllocateNative*Mat() from dangerous copy semantic.
The ability of initializing the created native Mat from the content
to which the provided *cArray pointer points to (if it is non-NULL)
was not only never used in any module - it is a quite useless feature,
but its unexpected semantic is highly dangerous and allows to easily
introduce crash-bugs!

Get rid of this awful and dangerous misfeature.