PsychPortAudio: Optimize audio data buffer exchange on Python.
Use the PsychUseCMemoryLayoutIfOptimal(TRUE) function to request
C-Memory layout for data exchanges if that is considered optimal
on a scripting environment, and adapt processign accordingly.

On Python C-layout is optimal, as it is the default layout used
by NumPy for >= 2D arrays, so we use appropriate help texts to
ask for matrices with 1 row == 1 sample frame, 1 column == 1 channel,
instead of the Matlab/Octave format with 1 row == 1 channel and
1 column == 1 sample frame. This way we get data in optimal row-
major format.

Code changes:

1. Adapt help texts/synopsis accordingly on Python.

2. Expect switched row vs. column semantics and also take it into
   account for error handling / error messages.

3. On the 'GetAudioData' output side, switch row/column accordingly
   to return C optimized.

4. 'GetAudioData' returns data in float32 aka single() aks float
   format by default, unless requested by usercode otherwise. On
   Matlab/Octave we default to double format (float64) for backwards
   compatibility.

Update ppatest.py accordingly, although the current code is a hack,
transposing matrices after the fact, but this will do for testing.

-> On a 2010 MacBookPro this provices a speedup of ~ factor 3 for
   a 48000 samples matrix, so it can be quite significant. Using
   float instead of double gives another speed bump.