Screen: Allow to skip the wait-out-of-vblank workaround before Flips.
A new ConserveVRAM setting kPsychSkipOutOfVblankWait (2^29) allows to
skip the wait for being more than 2 msecs from last swap away, before
issuing a new swap request. This to enable use of > 500 Hz displays.

This is a bit crude. A better solution would be to adapt the minimum
wait time to be == the duration of the vblank. However, this is not the
right time for clever experiments…