PsychPortAudio: Switch priority between WDM/KS and WASAPI on Windows.
WASAPI is supposed to be usable since Windows-7, improved in
Windows-8.1 and pretty good for pro-audio use since current
Windows-10, whereas WDM/KS was the best Windows had to offer
in XP/Vista days.

Change order for low-latency api's to prefer WASAPI over WDM/KS.

When we are at it, delete some dead code as well.