PsychPortAudio: Upgrade portaudio 64-Bit binaries for Windows to v19.6.0+
Replace our custom made binaries for 64-Bit Windows by the upstream
master build, corresponding to v19.6.0+:

- MinGW64 portaudio_x64.lib for use with 64-Bit Octave.
- portaudio_x64.lib for use with 64-Bit Matlab.
- portaudio_x64.dll for runtime use with 64-Bit Matlab/Octave.

This build has Windows MME, WDM-KS and WASAPI enabled, but
DirectSound and ASIO compile-time disabled/removed.

It was built on Windows-10 with MSVC 2017, against the Win8.1 SDK.
Running it requires the "Microsoft Visual C++ 2015 redistributable
Update 3". At least when testing on Windows-7 it was neccessary to
install it.

Also remove the 32-Bit libs of portaudio, given that we are 64-Bit
only since a while.