PsychPortAudio: Add low-latency tuning and channel mapping for WASAPI.
Try to translate our 'selectchannels' parameter in 'Open' to some
meaningful output channelMask for WASAPI, to give users some control
over which sound channels go to which speakers/outputs. The same
mapping would also apply to the other Windows api's, ie. WDMKS,
DirectSound and MME, so implement common code path, and some empty
bits for addressing the other API's.

Implement optimized setup for 'reqlatencyclass' > 1.

We try to get raw audio passthrough, without any DSP effects,
however disabling AOP DSP effects is only supported on Windows 8.1
and later.

At reqlatencyclass 2+ we request exclusive mode access for the audio
device, which also implies event driven audio-callbacks, and this
combo is supposed to be the best-timing, lowest latency opmode for
Windows WASAPI. Especially on Windows-10.

At reqlatencyclass 3+ we additionally enforce use of the audio
format we as client use, hopefully improving latency further,
but if this actually helps without compatibility issues needs
to be found out. We can only do this on Windows-10, on earlier
OS'es we fail silently doing this, as earlier Windows versions
don't support it.

At reqlatencyclass 4+ we fail if we are not on Windows-10, as
only Windows-10 supports reqlatencyclass 3 and higher.

All in all, optimal use of WASAPI really requires Windows-10
for optimal performance, timing and low-latency, and even for
lesser demands Windows 8.1 is strongly recommended.

See following MSDN article for reference wrt. the state of the art
in "Windows land":

https://docs.microsoft.com/en-gb/windows-hardware/drivers/audio/low-latency-audio

-> We now need to detect if we are running on Windows-10, so extend
   PsychTimeGlue.c to allow to use the detection code with PsychPortAudio.
   This requires linking against advapi32.lib, which is set up for the
   windowmakeit... script for Octave and Matlab, but tbd. for other
   build scripts.

-> The MSDN docs also suggest use of a special MS-Windows provided
   driver to achieve lowest possible latency on HDA compatible hardware,
   as good as a roundtrip latency of about ~ 10 msecs best case.

-> Test of correctness and actual expected performance improvement
   pending.