setup.py: Shuffle things around for testing.
No longer delay-load portaudio_x64.dll, but link normally. Currently it
does not have a benefit doing otherwise.

Shuffle around where files are going. Now we host all .py files and
required runtime libraries (only needed on Windows) in
Psychtoolbox/PsychPython for the moment, ie. we include libusb-1.0.dll
and portaudio_x64.dll there. These are just copies (!) of the same dll's
from PsychContributed/x64 and PsychSound, because i fought hard with
fracking setup tools for hours, googled myself to death and lost in the
end. Apparently creating an easy way to distribute all dependencies of a
Python package is another thing that is done rather poorly by default. I
so hope i'm just wrong and ignorant in this case...

The psychtoolboxclassic.py init routine was replaced by a __init__.py
routine for the package, otherwise with the same content.

This is probably not the final word on this, but will do for the moment.