PsychHID: Allow to restrict KbQueues input to specific screens or windows.
PsychHID('KbQueueCreate') now accepts a new optional parameter 'windowHandle'.
This parameter is silently ignored on MS-Window, macOS and Linux under Wayland.

On Linux/X11 the optional windowHandle can be the screenId+1 of a X-Screen,
e.g., 1 for X-Screen 0, 2 for X-Screen 1 ..., or the X-Window handle of a
PTB onscreen window. Such a X-Window handle can be gained via a new
Screen('GetWindowInfo') subfield.

If windowHandle is zero, input is requested for all X-Screens, as in the past.

If windowHandle is screenId+1, input is only requested for X-Screen screenId.

If windowHandle is a X-Window handle, input is only requested for that onscreen
window.

-> This functionality is mostly useful for use with touchscreen input, to
   hopefully avoid or workaround various oddities wrt. mouse pointer emulation,
   coordinate remapping or multi-x-screen ops. The jury is out on this...

KbQueueCreate() now takes an optional PTB screenId or onscreen window handle
and translates it into a proper handle for PsychHID('KbQueueCreate').

TouchQueueCreate() now actually passes the given windowHandle to
KbQueueCreate -> PsychHID('KbQueueCreate').

This is another attempt at fixing odd touchscreen input problems,
especially on a multi-X-Screen setup or multi-touchscreen setup,
as reported in forum messages.

Symptoms on multi-X-Screen are currently:

- 1st touchpoint having a wrong mapping to screen, apparently due to
  some interference of mouse pointer emulation?

- 2nd touch screen "stealing" the 1st touchpoint of 1st touch screen.

- Other weirdness.

-> This needs actual testing by users, i don't have a real touchscreen atm.