Screen: Try to fix MakeTexture for double -> uint8 texture creation.
Compute offset to add when converting double input matrices into uint8
textures, so that the uint8 texture data matches how the hardware
converts floating point color values to fixed-point integer color
values. When we convert a double float matrix into uint8 textures, we
want to apply the same conversion behavior that the hw applies when
drawing "floating point color" primitives into a fixed point
framebuffer. This is important, e.g., for a uint8 texture created from
double 0.5 luminance values to draw the same color value into a 8 bpc
framebuffer, as a framebuffer clear command, drawdots command etc. with
normalized color 0.5 does.

See what happens if this goes wrong (as of Psychtoolbox 3.0.14) in
forum message #23300.

The original OpenGL 1.0 spec says hw should convert float -> fixed with
round-to-nearest behaviour (Section 2.12.9). The OpenGL 4.5 spec leaves
it up to the GL implementation if it wants to truncate, round, ceil up
or down to the closest larger or smaller integer etc., but recommends
round-to-nearest (Section 2.3.5.2). In practice, tested AMD and NVidia
hw on Linux and macOS truncates, whereas some tested Intel hw on Linux
and macOS rounds to nearest, so we apparently can't rely on defined
behaviour across hardware vendors and therefore detect what the given
hw does during startup and adapt accordingly when converting double
data to uint8 textures.

Note that an easier route would be to pass the input data for 8 bpc
textures as GL_FLOAT and leave the job to the driver. That however
would quadruple the memory and cache footprint and memory bandwidth
required for ‘MakeTexture’ and require additional cpu computation in
the driver, therefore could cause a substantial performance loss which
we want to avoid.