Python: Get rid of redundant copies of dll's in PsychPython/
On Windows, we only need copies of portaudio_x64.dll and
libusb-1.0.dll to be present during execution of setup()
by setup.py. Avoid carrying redundant copies for that purpose.

Instead, setup.py on Windows copies those dll's from their
actual locations in PsychSound/ and PsychContributed/ to
PsychPython/, executes setup(), then deletes them from
PsychPython again.

This saves space in the repo, but also ensures that the
dll's delivered to PTB for Python are always in sync with
those used on Octave and Matlab.

Btw. yes this hack is kind of lame, but none of distutils
provided distribution methods for files allows for the
needed way of bundling afaik. data_files parameter only
allows copying into absolute dirs or relative dirs to the
base directory...