PsychPortAudio: Add workaround for Portaudio audio capture bug on Linux+ALSA.
Does anybody still test software before (or at least after) release?!

All shipping versions of Portaudio since over 10 years (!) have a bug in
their half-duplex "audio capture only" mode on Linux+ALSA, which cause
awful recording artifacts, at least as tested on multiple Intel HDA
sound chips.

Add the following workaround, to allow us to dynamically link with distro
provided broken libportaudio, "fixing" the bug:

The bug triggers in pure half-duplex audio capture
mode (== no playback requested) and consists of erroneously switching
the hostBufferSizeMode of the PaBufferProcessor for capture streams from
paUtilFixedHostBufferSize to paUtilBoundedHostBufferSize, causing
misconversion of captured audio somewhere in the bufferProcessor.

The offending (buggy) if () conditional is at the end of the function
PaAlsaStream_DetermineFramesPerBuffer() inside pa_linux_alsa.c. It looks
like this: if( !self->playback.canMmap || !accurate ) ..., when it should look
like this: if((!self->playback.canMmap && self->playback.pcm) || !accurate ) ...
in order to prevent the invalid !self->playback.canMmap check, which always
evaluates to "true" if playback is not actually used in pure capture mode.

We can't prevent the mis-setting of hostBufferSizeMode to paUtilBoundedHostBufferSize,
but we can reset it to the correct paUtilFixedHostBufferSize after Pa_OpenStream()
and before actual start of capture operations, by hacking ourselves into the internal
data structure of the PaBufferProcessor associated with the audio stream.

We don't apply the fix if usercode specified a buffersize other than
the default paFramesPerBufferUnspecified -- in such a case, switching
to a different hostBufferSizeMode may be required...

Btw. testing on Ubuntu 18.04.1 LTS showed that at least Ubuntu, but possibly
Debian as well, ships a patched portaudio version, which changed some data
structures that we hack into, so now we already need special-case handling for
those Ubuntu style variants. I hope that these won't be the only incompatible
variants, but this approach is highly fragile to any deviations of bundled
portaudio libraries from vanilla portaudio upstream.

May still need to go back to bundling our own statically or dynamically linked
versions of portaudio for Linux if this proves unreliable, or if more bugs
show up in portaudio 19.6...

-> Test on other distros than Ubuntu 18.04.1 LTS 64-Bit Intel pending.