Delete most mex files for 32-Bit Octave + Intel on Linux.
Those are unmaintained since a long time, and no longer compatible
with M-Files in the distribution.

People who want to stick to 32-Bit Intel can simply use the up-to-date
builds of NeuroDebian and Debian/Ubuntu et al. builds.

Keep Datapixx.mex and Eyelink.mex, as NeuroDebian and other distros
do not ship their own builds of these, and this old ones should be
still ok.