CreateProceduralSmoothedApertureSineGrating(): Clarify interactions with alpha blending.
Add definitions of the formulas to the help text which explain better how the
radius and sigma and useAlpha parameters influence the calculation of color
and alpha values, and how this interacts with alpha blending modes.

For clarity, in ProceduralSmoothedApertureSineGratingDemo.m
clarify interaction of alpha-blending with useAlpha and radius=inf,
and the sine modulation that is conditionally also applied to the
alpha-channel for radius=inf or useAlpha=false.

Also add a bit of code to SineGratingSmoothedApertureShader.frag.txt
to make sure that modulateColor.a aka globalAlpha from 'DrawTexture' and
the Offset color alpha gets properly applied in the useAlpha=true case.

-> This should not change the behavior of existing code with radius < inf
   and useAlpha=true, unless somebody does something non-sensical with
   modulateColor, globalAlpha or backGroundColorOffset. Iow. the change
   should be backwards compatible with reasonable code.

The whole interaction of alpha channels with alpha-blending is a complex
topic that probably should have more detailed documentation, as it seems
to be unintuitive to many people. This is just a stop-gap to make stuff
moderately less confusing to people which didn't read the ECVP2013 PTBTutorial's
section about alpha blending in our PsychDocumentation folder, or further OpenGL
documentation about the topic. While our demos "as they are" show safe use
of the shaders, if somebody deviates from them esp. wrt. alpha blending settings,
the results are often surprising to them.

For more background and ongoing discussions see GitHub issue #505