Check-in of initial skeleton PsychOculusVRCore1 driver.
The driver source files are white-space/indentation/formatting cleaned
and properly renamed files of the old (v0.5 SDK based) cross-platform
driver that was/is used with the v0.5 SDK/Runtime on Linux/OSX/Windows.
Some clearly unneeded code already stripped.

There is still lots of dead code in this, which will need to be stripped
during ongoing development, and none of the SDK API calls was converted
to the new 1.0+ SDK, so none of this will build atm., but it defines
the baseline skeleton for iterating.