PsychOculusVRCore1: Add proof-of-concept OpenGL rendering and display support.
New methods 'CreateRenderTextureChain' and 'GetNextTextureHandle' allow
to create a texture swap chain for submitting frames to the VR
compositor, and to get the OpenGL texture handle of the next
render-to-texture target texture in the chain.

Function 'EndFrameTiming' contains minimal code to commit such a freshly
rendered-to texture to the swap chain, and to submit a frame to the VR
compositor for display asap.

-> Minimal infrastructure for testing rendering+display path, work in
progress.