PsychOculusVR1: Implement imaging pipeline setup and Flip-Processing.
Setup imaging pipeline processing hooks to process our per-Flip
PreSwapbuffersOperations and shutdown callbacks.

Implement the texture processing and submission to drive the HMD.

-> This version gives the "first light" with a mono- or stereo
imagestream presented on the HMD that is somewhat sensible, although not
yet completely correct. The mechanics of driving the
rendering-warping-display works, but there are mistakes:

- The image display is wrongly interlocked with head tracking. Head
movements right/left move the rendered image out of the view in the HMD.
Need to attach our layers in a different way i guess.

- While rendering of the stereo image pairs more or less works, 3D
rendering in OpenGl (e.g., VRHMDDemo1) where the OpenGL camera is driven
by the projection/modelview matrices from the driver seems to be
slightly wrong, and there seems to be a left-right confusion for the
rendered views, so all stereo disparities are reversed and in
cue-conflict with other 3D/Depth cues.

-> Needs more work, but at least one milestone of seeing something not
totally wrong is reached. Backup commit.