PsychOculusVR1: Fix left vs. right eye confusion during 3D rendering.
Contrary to what the v0.5 SDK did, the stable 1.x SDK now provides
HmdToEyeOffsets which don't have their signs switched: The
HmdToEyeViewOffsetLeft actually has a negative x-axis component,
shifting the left eye camera to the left of the HMD's tracked center,
and the HmdToEyeViewOffsetRight has a positive x-axis component,
shifting the right eye camera to the right of the HMD's tracked center.

Remove the -1 multiply for sign switching, which is no longer needed,
but harmful, and caused the left- vs. right eye confusion during
rendering. With this fix, VRHMDDemo1 and SuperShapeDemo now render
correct stereo pairs for proper display.