PsychOculusVR1: Clear compositor textures to white at startup.
Iterate through all textures in the swap chain and fill them with a
constant 1.0 initial value for RGBA. This is needed if PTB's imaging
pipeline has some image processing ops enabled which do not fully
overwrite the finalizedFBO output framebuffers, and thereby would leave
parts of the returned compositor textures untouched/uninitialized. An
example is use of the panel-fitter with a 'Centered' fitting mode for a
framebuffer smaller than the size of the compositor output textures.
This would center output in the textures, but leave borders untouched.

-> We choose white as initial clear color, as any other values than
constant 1.0 i tried do not work.