PsychOculusVR1: Add new 'DebugDisplay' mode.
Specifying the keyword 'DebugDisplay' as part of the 'BasicRequirements'
parameter for 'SetupRenderingParameters' or 'AutoSetupHMD' will show the
PTB onscreen window and display a fullscreen display of the content
shown on the HMD, by sourcing compositor results from an auto-updating
mirror texture. This allows convenient debugging without the HMD on the
head.

Omitting the 'DebugOutput' keyword disabled the debug display and hides
the PTB onscreen window from actually showing on the desktop at all, as
it would not display anything useful anyway.