PsychOculusVR1: Split 'flip' processing into texture commit and frame submit.
We call fast-path 0 from the finalizerBlitChain as part of preflip
operations to detach ready textures and commit them into the texture
swap chains.
We call fast-pasth 1 from the PreSwapbuffersOperations chain for
submitting the frame to the VR compositor, waiting for Present
completion/mirror texture draw/timestamp injection.

-> Works, but flipping is not reliably throttled to Present completion,
and timestamping not yet implemented.