PsychOculusVRCore1: Add compositor timeout prevention. First iteration.
We have two basic modes of operation, VR3DTracked and non-tracked.

In VR3DTracked mode we use the HMD as its creators intended it to
be used - to present dynamically changing VR 3D scenes, driven by
head tracking information, with minimal "motion to photon" latency.
In this case the M-File script will run in a tight loop where it
gets recent head pose data from tracking, computes updated camera
and scene data, renders a new frame and Screen('Flip', window);
presents it asap on the HMD. In this mode we want updates at HMD
refresh rate, and we want the VR compositor to trigger timeout
handling and safeties if usercode doesn't Flip within the timeout
interval of about 0.1 seconds. The user is expected to perform
substantial head or body movements and would get motion sick or
bump into stuff otherwise.

In non-tracked mode, our users code (ab)uses the HMD as a strap-on
high quality mono monitor or stereoscopic 3D display without any
closed loop feedback from head-tracking to rendering, in fact most
likely the subject is seated in a chair, not moving head or body
much at all. In this case the usercode may present stereo movies,
stereo video, 3d static/dynamic stimuli, or binocular rivalry
imagery. In this case there may pass substantial time between
stimulus updates by use of Screen('Flip', window, when); 'when'
target presentation times in excess of the ~ 0.1 secs compositor
timeout. This would trigger timeout handling and disrupt the
visual stimulation paradigm!

Prevent timeout by use of a "timeout prevention idle thread":

We run a separate background processing thread which wakes up
once per HMD video refresh, checks if a frame has not been presented
to the compositor for almost a full timeout duration, and if so, it
may decide to ovr_SubmitFrame() the last queued frame (the last
one committed to the texture swapchains) again, to reset the
timeout timer of the compositor.

The compositor and 'Present' method of the driver keep track of
the last tine a present was executed, and thereby of the current
timeout deadline. The thread tries to take scheduled future
presentation time into account if it knows about it, ie. the
'when' time passed by usercode to Screen('Flip'). If the timetamp
suggests as present of new rendered content is about to happen, it
will do nothing but wait out how things evolve. If it decides the
'when' time would be past the timeout deadline, or doesn't know the
'when' time, it will force (re-)present old content when it is closer
than 1 HMD frame duration to the timeout.

The timeout is currently estimated to be 10 HMD refresh durations,
but that is just guesswork.