PsychOculusVRCore1: Partial fixup of multi-threaded stimonset timestamping.
This introduces the use of condition variables to avoid spin-waits when
the mainthread needs to wait for results from the presenterThread. Also
the PresentExec function now calls ovr_GetPerfStats() and hands data
lock-protected to the mainthread as consumer. However, so far this is
mostly broken, causing either hangs or degraded performance.

-> Backup commit. All demos/tests need to disable "precise" timestamping
to work ok at this point.