PsychOculusVR: Assign individual field of view per eye, not globally for both eyes.
While the Rift DK1 and Rift DK2 apparently had an almost symmetrical
field of view for both eyes (only 0.896 degrees horizontal difference
under the v0.5.0.1 SDK and old PsychOculusVR driver, and only 1.9315
degrees horizontal difference under the v1.14 SDK and new PsychOculusVR1
driver), ie. almost identical FOV and also horizontally symmetric, this
is no longer the case for the Rift CV1. It uses an asymmetric FOV for
each eye, much narrower towards the horizontal direction pointing to the
nose (which is a natural occluder), and thereby also asymmetric for the
two eyes, with a total difference of 53 degrees in horizontal angles
(abs(left - right fov)). In the past we always used the right-eye FOV
for both eyes, which worked without any perceptible visual
artifacts/distortions/shifts, but now causes unwanted shifts in
rendering on the CV1 with this large difference between left and right
borders of the frustum and correspondingly between the two eyes.

Therefore, ask the Oculus runtime for recommended separate left-eye and
right-eye FOV and use them.

-> New problem: Now all 3D rendering user scripts must use per-eye
projection matrices. Most of our demos don't do this so far, so they
need to be fixed up.
-> New problem: Rendering of Monoscopic/Stereoscopic views (without
tracking) is utterly wrong, as standard Screen() 2D drawing uses
identical projection/modelview matrices and viewports for both eyes in a
stereo config. Need to find a backwards compatibility fix for that use
case as well.