PsychOculusVR1: Apply symmetric and identical field of view in 2D non-tracked display modes.
If 'basicTask' == 'Monoscopic' or 'Stereoscopic' is selected, use an
identical field of view for left and right eye view, and make it
symmetric in the horizontal direction, as the minimum of left- and
right- view angle. This will avoid the need for 2D drawing and display
scripts to use special projection matrices, making PsychOculusVR1
backwards compatible with PsychOculusVR and PsychVRHMD. Usercode for 2D
display (e.g., ImagingStereoDemo, VRHMDDemo, ImagingStereoMoviePlayer)
can run unmodified on the Rift CV1.