PsychOculusVRCore1: Report sessionStatus a part of 'GetTrackingState'
Returns state.sessionStatus with ovr_GetSessionStatus() flags.

PsychOculusVR1('PrepareRender') returns this state.sessionStatus flags
which tell the caller if the HMD is present, active, we have control
over it, and it is actually strapped to the subjects head and thereby
active and displaying our content. Also reports failure conditions, UX
requests to quit the session, and UX requests to recenter tracking
origin.

'PrepareRender' automatically error() aborts a session if it detects the
'DisplayLost' fatal error condition.