PsychOculusVR[Core]1: Add subfunctions 'FloatsProperty' and 'StringProperty'
These allow to query and set float, float-array, and string properties.
'FloatsProperty' works with float and float-array properties.

-> While queries work nicely, as tested in VRInputStuffTest.m, setting
new values failed on every supported property with a "property is
read-only" error. This contradicts SDK 1.11 docs, but i can't find a
mistake in the code, and this happens both for string, float and
float-array props, both predefined ones, and new ones as defined in the
OVR_CAPI_Keys.h header file.

-> VRInputStuffTest.m slightly improved in formatting etc.