PsychOculusVRCore1: Return empty 'frameTiming' when low-precision timestamping.
If 'PresentFrame' is not used with high precision timestamping -- which is somewhat fragile at the moment -- we still must return an empty 'frameTiming' struct array as 1st optional return argument, as otherwise the calling scripting code will bomb out badly due to "1st return argument expected but not assigned" errors. This would not cause error-abort, but rendering errors and visual corruption.