PsychOculusVRCore1: Add 'valid' field to 'GetInputState'.
If the HMD is not on the users head, ie. invisible, then we lose
VR input focus and don't get meaningful controller input anymore
from ovr_GetInputState(). In this case we shouldn't error abort, but
just return an all-zero 'input' struct.

We have a new subfield "Valid" in the struct, which returns 1
if input was valid, 0 if we don't get input due to VR focus loss.
The "Valid" subfield was added to all drivers and PsychVRHMD,
so this is a field guaranteed to be always present.