DrawFormattedText/DrawFormattedText2: Suppress unicode warnings on Octave.
Octave can't char() cast unicode values > 255, which we handle.
It also prints a warning() about that fact on each invocation of
DrawFormatttedText[2](), which we so far didn't handle.

Add Octave specific code to disable warnings() while the function
executes, auto-reenable on normal or error exit. We can't suppress
the specific warning as it doesn't have a warning id assigned.

Tested on Octave 4.2.2, Matlab should have compatible warning()/onCleanup()
implementation, but we skip it as it is not needed on Matlab.