Screen/Linux: Another attempt at screen->displayhead mapping for AMD.
The new code that fixed the problem on DCE-8 with and without DC,
tested on Linux 4.15 iirc, still broke on DCE-11 Polaris when tested
with Linux 4.20-rc4.

Let's assume there's a difference in hw between DCE-8 and later
engines and use a new weird mix of old and new code-path for DCE-11
and DCE-10. Fixes Polaris DCE-11, i don't know about DCE-10 due to
lack of test hardware, but assume for the moment it behaves like
DCE-11.

Could also be that this has nothing to do with DCE version, but
with Linux kernel version, although i couldn't find any obvious
difference in gamma lut hw programming between 4.15 and 4.20-rc.
In that case we will be in for more debugging fun.